+++ lexicmap/cmd/seed-pos.go
@@ -49,6 +49,7 @@ var seedPosCmd = &cobra.Command{
 	Long: `Extract seed positions via reference names
 
 Attentions:
+  0. This command requires the index to be created with the flag --save-seed-pos in lexicmap index.
   1. Seed/K-mer positions (column pos) are 1-based.
      For reference genomes with multiple sequences, the sequences were
      concatenated to a single sequence with intervals of (k-1) N's.